requirements.txt
aiofiles==24.1.0
fastapi==0.115.12
uvicorn[standard]
sqlalchemy
pydantic==2.10.5
email-validator==2.2.0
pymysql==1.1.1
python-multipart==0.0.20
psutil==7.0.0
PyJWT==2.10.1
passlib==1.7.4
bcrypt==3.2.0
motor==3.7.0
pytest==8.3.5
minio==7.2.15
cryptography==39.0.1
itsdangerous==2.2.0
httpx==0.24.0
Authlib==1.3.0
pytz
redis>=4.5.0,<5.0.0
celery>=5.2.0,<5.5.0
aiohttp==3.11.18
weasyprint>=54.0
markdown2>=2.4.0
python-dotenv==1.0.1
python-jose==3.5.0


